+++ src/components/Tooltip/Readme.md
@@ -1,0 +1,21 @@
+```jsx
+<Tooltip position='top' content="My tooltip content. Lorem ipsum dolar set amet.">
+  <span>Top</span>
+</Tooltip>
+```
+
+```jsx
+<Tooltip smart content="My tooltip content. Lorem ipsum dolar set amet.">
+  <span>Smart Positioning</span>
+</Tooltip>
+```
+
+```jsx
+<Tooltip
+  animation='fade'
+  position='bottom'
+  content="My tooltip content. Lorem ipsum dolar set amet."
+>
+  <span>Bottom</span>
+</Tooltip>
+```
